Added notifications to 3 pages
imported the notifications component into the three other pages.

If this commit is causing issues feel free to discard it and add the changes manually later.

All it requires is that the code is within a main element of class content. Check the commit for the code snippet to add