SSE Overhaul
Changed the setup of server side events so that there is a constant loop running on the server once the server is started to generate data. This only generates 100 points of data before it starts deleting old data so as not to overflow the database. This number can be changed at the top of scheduler.js. The server side events keeps a list of active connections and updates all active connections with the new data once availible.

The server saves the generated data to the database under the AI_Logs model.